requirements.txt
python-telegram-bot>=21.0
openai>=1.0
psycopg2-binary>=2.9
pytest>=7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
httpx>=0.27.0
argon2-cffi>=23.1.0
sentry-sdk[fastapi]>=2.0.0
